moved note methods tests from note-test to note-method-test
